TIP-810: Move Reset index command in ElasticSearchBundle
